Last Comit of Project.

+++ Entities/Computer.cs
@@ -19,7 +19,7 @@ namespace Entities
         public string? OperatingSystem { get; set; }
 
         public string? Ram { get; set; }
-        public float ScreenSize { get; set; }
+        public int ScreenSize { get; set; }
 
 
     }
